Move creatable targeted notification types to /models
