+++ src/client.spec.ts
@@ -16,7 +16,6 @@ import {
 	CreatableSavedQuery,
 	CreatableScheduledQuery,
 	CreatableScheduledScript,
-	CreatableTargetedNotification,
 	CreatableTemplate,
 	CreatableUser,
 	DashboardsFilter,
@@ -34,7 +33,6 @@ import {
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
-	TargetedNotificationTargetType,
 	UpdatableNotification,
 	UpdatablePlaybook,
 	UpdatableResource,
@@ -60,6 +58,7 @@ import {
 	CreatableJSONEntry,
 	CreatableMacro,
 	CreatableMultiLineEntry,
+	CreatableTargetedNotification,
 	Dashboard,
 	FileMetadata,
 	Group,
@@ -81,6 +80,7 @@ import {
 	Search,
 	Search2,
 	SearchModule,
+	TargetedNotificationTargetType,
 	Template,
 	UpdatableActionable,
 	UpdatableAutoExtractor,
