+++ src/functions/notifications/create-one-targeted-notification.ts
@@ -6,14 +6,18 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { toRawNumericID } from '../../value-objects';
+import {
+	CreatableTargetedNotification,
+	CreatableTargetedNotificationByTargetType,
+	TargetedNotificationTargetType,
+	toRawCreatableTargetedNotification,
+} from '../../models';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
 	buildURL,
 	fetch,
 	HTTPRequestOptions,
-	omitUndefinedShallow,
 	parseJSONResponse,
 	URLOptions,
 } from '../utils';
@@ -31,7 +35,7 @@ export const makeCreateOneTargetedNotification = (makerOptions: APIFunctionMaker
 
 			const baseRequestOptions: HTTPRequestOptions = {
 				headers: { Authorization: authToken ? `Bearer ${authToken}` : undefined },
-				body: JSON.stringify(toCreatableRaw(_creatable)),
+				body: JSON.stringify(toRawCreatableTargetedNotification(_creatable)),
 			};
 			const req = buildHTTPRequest(baseRequestOptions);
 
@@ -60,110 +64,5 @@ const _buildURL = (creatable: CreatableTargetedNotification, baseOptions: URLOpt
 			// const templatePath = '/api/notifications/targeted/group/{groupID}';
 			return buildURL(templatePath, baseOptions, { pathParams: { groupID: creatable.groupID } });
 		}
-	}
-};
-
-export type TargetedNotificationTargetType = CreatableTargetedNotification['targetType'];
-type CreatableTargetedNotificationByTargetType<
-	TargetType extends TargetedNotificationTargetType
-> = TargetType extends 'myself'
-	? CreatableMyselfTargetedNotification
-	: TargetType extends 'group'
-	? CreatableGroupTargetedNotification
-	: TargetType extends 'user'
-	? CreatableUserTargetedNotification
-	: never;
-
-export type CreatableTargetedNotification =
-	| CreatableMyselfTargetedNotification
-	| CreatableGroupTargetedNotification
-	| CreatableUserTargetedNotification;
-
-export interface CreatableMyselfTargetedNotification extends CreatableBaseTargetedNotification {
-	targetType: 'myself';
-}
-
-export interface CreatableGroupTargetedNotification extends CreatableBaseTargetedNotification {
-	targetType: 'group';
-
-	/**
-	 * ID of the group targeted by the notification
-	 */
-	groupID: string;
-}
-
-export interface CreatableUserTargetedNotification extends CreatableBaseTargetedNotification {
-	targetType: 'user';
-
-	/**
-	 * ID of the user targeted by the notification
-	 */
-	userID: string;
-}
-
-interface CreatableBaseTargetedNotification {
-	message: string;
-	customID?: string;
-
-	sentDate?: string; // Timestamp eg. '2019-04-22T21:44:01.776942432Z'
-	expirationDate?: string; // Timestamp eg. '2019-04-23T03:44:01.776918756-06:00'
-	ignoreUntilDate?: string; // Timestamp eg. '0001-01-01T00:00:00Z'
-}
-
-type RawCreatableTargetedNotification =
-	| RawCreatableMyselfTargetedNotification
-	| RawCreatableGroupTargetedNotification
-	| RawCreatableUserTargetedNotification;
-
-interface RawCreatableMyselfTargetedNotification extends RawCreatableBaseTargetedNotification {
-	UID: 0;
-	GID: 0;
-}
-
-interface RawCreatableGroupTargetedNotification extends RawCreatableBaseTargetedNotification {
-	UID: 0;
-	GID: number; // 0 is undefined
-}
-
-interface RawCreatableUserTargetedNotification extends RawCreatableBaseTargetedNotification {
-	UID: number; // 0 is undefined
-	GID: 0;
-}
-
-interface RawCreatableBaseTargetedNotification {
-	Type?: number;
-	Broadcast: false;
-	Sent?: string; // Timestamp eg. '2019-04-22T21:44:01.776942432Z'
-	Expires?: string; // Timestamp eg. '2019-04-23T03:44:01.776918756-06:00'
-	IgnoreUntil?: string; // Timestamp eg. '0001-01-01T00:00:00Z'
-	Msg: string;
-}
-
-const toCreatableRaw = (creatable: CreatableTargetedNotification): RawCreatableTargetedNotification => {
-	const base: RawCreatableBaseTargetedNotification = {
-		Msg: creatable.message,
-		Type: creatable.customID === undefined ? undefined : toRawNumericID(creatable.customID),
-		Broadcast: false,
-
-		Sent: creatable.sentDate,
-		Expires: creatable.expirationDate,
-		IgnoreUntil: creatable.ignoreUntilDate,
-	};
-
-	switch (creatable.targetType) {
-		case 'myself':
-			return omitUndefinedShallow<RawCreatableMyselfTargetedNotification>({ ...base, GID: 0, UID: 0 });
-		case 'group':
-			return omitUndefinedShallow<RawCreatableGroupTargetedNotification>({
-				...base,
-				UID: 0,
-				GID: toRawNumericID(creatable.groupID),
-			});
-		case 'user':
-			return omitUndefinedShallow<RawCreatableUserTargetedNotification>({
-				...base,
-				GID: 0,
-				UID: toRawNumericID(creatable.userID),
-			});
 	}
 };
