+++ src/models/notification/index.ts
@@ -8,10 +8,14 @@
 
 export * from './broadcasted-notification';
 export * from './creatable-broadcasted-notification';
+export * from './creatable-targeted-notification';
+export * from './creatable-targeted-notification-by-target-type';
 export * from './notification';
 export * from './raw-broadcasted-notification';
 export * from './raw-creatable-broadcasted-notification';
+export * from './raw-creatable-targeted-notification';
 export * from './raw-notification';
 export * from './raw-targeted-notification';
 export * from './targeted-notification';
 export * from './to-raw-creatable-broadcasted-notification';
+export * from './to-raw-creatable-targeted-notification';
